Slettet PdfKvitteringService og la på token på testen
