--- src/main/kotlin/no/nav/familie/ef/mottak/service/PdfSøknadService.kt
@@ -1,60 +1,0 @@
-package no.nav.familie.ef.mottak.service
-
-import no.nav.familie.ef.mottak.api.dto.Kvittering
-import no.nav.familie.ef.mottak.integration.FamilieDokumentClient
-import no.nav.familie.ef.mottak.mapper.SøknadMapper
-import no.nav.familie.ef.mottak.repository.SøknadRepository
-import no.nav.familie.ef.mottak.repository.VedleggRepository
-import no.nav.familie.ef.mottak.repository.domain.EncryptedFile
-import no.nav.familie.ef.mottak.repository.domain.Søknad
-import no.nav.familie.ef.mottak.repository.domain.Vedlegg
-import no.nav.familie.kontrakter.ef.søknad.SøknadMedVedlegg
-import no.nav.familie.kontrakter.ef.søknad.SøknadOvergangsstønad
-import org.slf4j.LoggerFactory
-import org.springframework.stereotype.Service
-import org.springframework.transaction.annotation.Transactional
-import java.util.UUID
-
-@Service
-@Transactional
-class PdfSøknadService(
-    private val søknadRepository: SøknadRepository,
-    private val vedleggRepository: VedleggRepository,
-    private val dokumentClient: FamilieDokumentClient,
-    private val taskProsesseringService: TaskProsesseringService,
-    private val pdfKvitteringService: PdfKvitteringService,
-) {
-    private val logger = LoggerFactory.getLogger(this::class.java)
-
-    @Transactional
-    fun mottaOvergangsstønad(søknad: SøknadMedVedlegg<SøknadOvergangsstønad>): Kvittering {
-        val søknadDb = SøknadMapper.fromDto(søknad.søknad, true)
-        val vedlegg = mapVedlegg(søknadDb.id, søknad.vedlegg)
-        return motta(søknadDb, vedlegg)
-    }
-
-    private fun mapVedlegg(
-        søknadDbId: String,
-        vedleggMetadata: List<no.nav.familie.kontrakter.ef.søknad.Vedlegg>,
-    ): List<Vedlegg> =
-        vedleggMetadata.map {
-            Vedlegg(
-                id = UUID.fromString(it.id),
-                søknadId = søknadDbId,
-                navn = it.navn,
-                tittel = it.tittel,
-                innhold = EncryptedFile(dokumentClient.hentVedlegg(it.id)),
-            )
-        }
-
-    private fun motta(
-        søknadDb: Søknad,
-        vedlegg: List<Vedlegg>,
-    ): Kvittering {
-        val lagretSkjema = søknadRepository.insert(søknadDb)
-        vedleggRepository.insertAll(vedlegg)
-        taskProsesseringService.startPdfKvitteringTaskProsessering(søknadDb)
-        logger.info("Mottatt pdf-søknad med id ${lagretSkjema.id}")
-        return Kvittering(lagretSkjema.id, "Pdf-søknad lagret med id ${lagretSkjema.id} er registrert mottatt.")
-    }
-}
